Create predict_review_2.py
In this program, apart from just making the predictions like in the predict_review.py program, we shall also plot the learning curve of the Naive Bayes model. We shall create another separate function to just plot the learning curve in the next version. Note, this code is a modification of the code by Adria Luz which can be found here:
https://gist.github.com/adrialuzllompart/c916c4ce3782a98ab5c92fe82ce0d293#file-plot_learning_curves-py